requirements.txt
pillow
sqlalchemy
fastapi
uvicorn
requests
adafruit-circuitpython-epd
